pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "dotenv>=0.9.9",
    "fastapi>=0.124.2",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pyjwt>=2.10.1",
    "pytest-xdist>=3.8.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "structlog>=25.5.0",
    "uvicorn>=0.38.0",
]


[dependency-groups]
dev = [
    "anyio>=4.12.0",
    "basedpyright>=1.36.1",
    "pytest>=9.0.2",
    "pytest-xdist>=3.8.0",
    "ruff>=0.14.8",
]

[tool.basedpyright]
reportAny=false
reportExplicitAny=false

[tool.pytest.ini_options]
anyio_mode = "auto"
